src/requirements.txt
Django
djangorestframework
django-filter
django-cors-headers
drf-spectacular
drf-yasg
psycopg2-binary
markdown
# Testing dependencies
factory-boy
faker
PyJWT[crypto]==2.8.0
cryptography==41.0.7
requests==2.31.0
djangoviz
setuptools


